v5.5.6 Beta #23
*Configuration for machine energy usage.
*Bumped default Theoretical Elementizer energy usage.
*Cleaned up code.
*Fixed UnsupportedOperationException in EnergyNetwork.

+++ src/minecraft/mekanism/common/TileEntityBasicMachine.java
@@ -72,7 +72,7 @@ public abstract class TileEntityBasicMachine extends TileEntityElectricBlock imp
 	 * @param ticksRequired - how many ticks it takes to run a cycle
 	 * @param maxEnergy - how much energy this machine can store
 	 */
-	public TileEntityBasicMachine(String soundPath, String name, String path, int perTick, int ticksRequired, double maxEnergy)
+	public TileEntityBasicMachine(String soundPath, String name, String path, double perTick, int ticksRequired, double maxEnergy)
 	{
 		super(name, maxEnergy);
 		ENERGY_PER_TICK = perTick;
